Feat: Add logic for displaying clicked components.
