+++ src/containers/Main.js
@@ -1,25 +1,16 @@
-import React, { useState } from 'react';
-import { useSelector, useDispatch } from 'react-redux';
+import React from 'react';
+import { useSelector } from 'react-redux';
 import PropTypes from 'prop-types';
 import {
   DashBoard, SideBar, Dashy,
 } from '../styles/StyledComponents';
 import Tittle from '../components/Sidebar/Title';
-import Slider from '../components/Dashboard/Slider/Slider';
 import Footer from '../components/Sidebar/Footer';
 import Pages from '../components/Sidebar/Pages';
+import dashboardHelper from '../helpers/dashboardHelper';
 
 const Main = ({ loggedInStatus }) => {
   const clicked = useSelector(state => state.pages.clicked);
-  const [click, setClick] = useState('');
-  const dispatch = useDispatch();
-
-  const images = [
-    'https://images.unsplash.com/photo-1449034446853-66c86144b0ad?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=2100&q=80',
-    'https://images.unsplash.com/photo-1470341223622-1019832be824?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=2288&q=80',
-    'https://images.unsplash.com/photo-1448630360428-65456885c650?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=2094&q=80',
-    'https://images.unsplash.com/photo-1534161308652-fdfcf10f62c4?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=2174&q=80',
-  ];
 
   return (
     <DashBoard>
@@ -30,7 +21,7 @@ const Main = ({ loggedInStatus }) => {
         <Footer />
       </SideBar>
       <Dashy>
-        <Slider slides={images} />
+        { dashboardHelper(clicked) }
       </Dashy>
     </DashBoard>
   );
