+++ src/helpers/pagesHelper.js
@@ -1,8 +1,17 @@
 import React from 'react';
+import { Holdable } from 'react-touch';
+import { pagesCreator } from '../Redux/actions/actions';
 import { PagesLink } from '../styles/StyledComponents';
 
-const pagesHelper = page => (
-  <PagesLink key={page}>{page}</PagesLink>
+const pagesHelper = (page, dispatch) => (
+  <PagesLink
+    onClick={() => {
+      dispatch(pagesCreator(page));
+    }}
+    key={page}
+  >
+    {page}
+  </PagesLink>
 );
 
 export default pagesHelper;
